change port on dev for front end so that it can auto reload
conatiner will run on port 5173 for front end. For development, it is
nice to have the front end auto reload to any changes. Using the
container front end would require restarting the container. Use a new
port for dev, but still be able to run the container to have a backend